% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.cox.R
\name{mgp.cox}
\alias{mgp.cox}
\title{Cox Proportional Hazard Regression Model}
\usage{
mgp.cox(x = x, y = y, time_var = time_var, covariable = NULL, data = data)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{time_var}{a character string indicating the time series variable.}

\item{covariable}{a character string indicating the covariates.}

\item{data}{a data.frame.}
}
\value{
fit_list
}
\description{
Cox Proportional Hazard Regression Model
}
