% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.glm_bi.R
\name{mgp.glm_bi}
\alias{mgp.glm_bi}
\title{Binomial Logistic Regression Analysis}
\usage{
mgp.glm_bi(x = x, y = y, covariable = NULL, data = data)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{data}{a data.frame.}
}
\value{
glm_list
}
\description{
Binomial Logistic Regression Analysis
}
