% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.glm_mult.R
\name{mgp.glm_mult}
\alias{mgp.glm_mult}
\title{Multinomial Logistic Regression Analysis}
\usage{
mgp.glm_mult(x = x, y = y, covariable = NULL, ref = NULL, data = data)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{ref}{a character string indicating the y reference level.}

\item{data}{a data.frame.}
}
\value{
fit_list
}
\description{
Multinomial Logistic Regression Analysis
}
