% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.glm_negb.R
\name{mgp.glm_negb}
\alias{mgp.glm_negb}
\title{Negative Binomial Regression Analysis}
\usage{
mgp.glm_negb(x = x, y = y, covariable = NULL, data = data)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{data}{a data.frame.}
}
\value{
glm_list
}
\description{
Negative Binomial Regression Analysis
}
