% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.glm_ord.R
\name{mgp.glm_ord}
\alias{mgp.glm_ord}
\title{Ordinal Logistic Regression Analysis}
\usage{
mgp.glm_ord(x = x, y = y, covariable = NULL, data = data)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{data}{a data.frame.}
}
\value{
fit_list
}
\description{
Ordinal Logistic Regression Analysis
}
