% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.glm_poisson.R
\name{mgp.glm_poisson}
\alias{mgp.glm_poisson}
\title{Poisson Regression Analysis}
\usage{
mgp.glm_poisson(x = x, y = y, covariable = NULL, data = data)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{data}{a data.frame.}
}
\value{
glm_list
}
\description{
Poisson Regression Analysis
}
