% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.lm.R
\name{mgp.lm}
\alias{mgp.lm}
\title{Linear Regression Analysis}
\usage{
mgp.lm(x = x, y = y, covariable = NULL, data = data)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{data}{a data.frame.}
}
\value{
lm_list
}
\description{
Linear Regression Analysis
}
