% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.mediate.R
\name{mgp.mediate}
\alias{mgp.mediate}
\title{mediation analysis}
\usage{
mgp.mediate(
  x = x,
  y = y,
  m = m,
  covariable = NULL,
  boot = F,
  sims = 100,
  out.model = c("glm", "lm"),
  med.model = c("glm", "lm"),
  data = data
)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{m}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{boot}{a logical value. Default is "F".}

\item{sims}{a number.}

\item{out.model}{glm or lm}

\item{med.model}{glm or lm}

\item{data}{a data.frame.}
}
\value{
med.list
}
\description{
mediation analysis
}
