% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.pcor.R
\name{mgp.pcor}
\alias{mgp.pcor}
\title{Partial / correlation analysis}
\usage{
mgp.pcor(
  x = x,
  y = y,
  cov = NULL,
  df = df,
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}

\item{cov}{a character string indicating the covariates.}

\item{df}{a data.frame.}

\item{method}{"pearson", "kendall", or "spearman".}
}
\value{
df_temp
}
\description{
Partial / correlation analysis
}
