% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.spline_cox.R
\name{mgp.spline_cox}
\alias{mgp.spline_cox}
\title{Restricted cubic spline for cox}
\usage{
mgp.spline_cox(
  y = y,
  x = x,
  covariable = NULL,
  spline = c(0.25, 0.5, 0.75),
  time_var = time_var,
  data = data
)
}
\arguments{
\item{y}{a character string.}

\item{x}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{spline}{The location of the knots. defaulted c(0.25,0.5,0.75).}

\item{time_var}{a character string indicating the time series variable.}

\item{data}{a data.frame.}
}
\value{
fit
}
\description{
Restricted cubic spline for cox
}
\examples{
ddist<-datadist(data)
options(datadist="ddist")
}
