% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgp.spline_lrm.R
\name{mgp.spline_lrm}
\alias{mgp.spline_lrm}
\title{Restricted cubic spline for logistic}
\usage{
mgp.spline_lrm(
  y = y,
  x = x,
  covariable = NULL,
  spline = c(0.25, 0.5, 0.75),
  data = data
)
}
\arguments{
\item{y}{a character string.}

\item{x}{a character string.}

\item{covariable}{a character string indicating the covariates.}

\item{spline}{The location of the knots. defaulted c(0.25,0.5,0.75).}

\item{data}{a data.frame.}
}
\value{
fit
}
\description{
Restricted cubic spline for logistic
}
